requirements.txt
altair==6.0.0
annotated-types==0.7.0
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
curl-cffi==0.15.0
frozendict==2.4.7
gitdb==4.0.12
gitpython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.194.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.11
jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
multitasking==0.0.12
narwhals==2.19.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
peewee==4.0.4
pillow==12.2.0
platformdirs==4.9.6
proto-plus==1.27.2
protobuf==5.29.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydeck==0.9.1
pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.37.0
requests==2.33.1
rich==14.3.3
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
yfinance==1.2.1
